feat(comment): 댓글 CRUD에 대한 유효성 검사 추가 및 예외 처리 강화

+++ src/main/java/com/sparta/scheduleapp/comment/service/CommentService.java
@@ -31,9 +31,13 @@ public class CommentService {
     private final UserRepository userRepository;
 
     @Transactional
-    public ResponseDto addComment(Long scheduleId, AddCommentRequestDto reqDto) {
-        User user = userRepository.findById(reqDto.getUserId()).orElseThrow(() -> new EntityNotFoundException("존재하지 않는 유저입니다."));
-        Schedule schedule = scheduleRepository.findByScheduleId(scheduleId).orElseThrow(() -> new EntityNotFoundException("존재하지 않는 일정입니다."));
+    public ResponseDto addComment(Long jwtUserId, Long scheduleId, AddCommentRequestDto reqDto) {
+        User user = userRepository.findById(jwtUserId).orElseThrow(
+                () -> new EntityNotFoundException("존재하지 않는 유저입니다.")
+        );
+        Schedule schedule = scheduleRepository.findByScheduleId(scheduleId).orElseThrow(
+                () -> new EntityNotFoundException("존재하지 않는 일정입니다.")
+        );
         Comment comment = new Comment(
                 schedule,
                 user,
@@ -45,16 +49,28 @@ public class CommentService {
     }
 
     public ResponseDto retrieveComment(Long scheduleId) {
-        Schedule schedule = scheduleRepository.findByScheduleId(scheduleId).orElseThrow(() -> new EntityNotFoundException("존재하지 않는 일정입니다."));
+        Schedule schedule = scheduleRepository.findByScheduleId(scheduleId).orElseThrow(
+                () -> new EntityNotFoundException("존재하지 않는 일정입니다.")
+        );
         List<Comment> comments = schedule.getComments();
 
         return new RetrieveCommentsResponseDto("댓글을 성공적으로 조회하였습니다.", comments);
     }
 
-    public ResponseDto editComment(Long scheduleId, Long commentId, EditCommentRequestDto reqDto) {
+    public ResponseDto editComment(Long jwtUserId, Long scheduleId, Long commentId, EditCommentRequestDto reqDto) {
+        User user = userRepository.findById(jwtUserId).orElseThrow(
+                () -> new EntityNotFoundException("존재하지 않는 유저입니다.")
+        );
+        scheduleRepository.findByScheduleId(scheduleId).orElseThrow(
+                () -> new EntityNotFoundException("존재하지 않는 일정입니다.")
+        );
+        Comment comment = commentRepository.getByCommentId(commentId).orElseThrow(
+                () -> new EntityNotFoundException("존재하지 않는 댓글입니다.")
+        );
 
-        scheduleRepository.findByScheduleId(scheduleId).orElseThrow(() -> new EntityNotFoundException("존재하지 않는 일정입니다."));
-        Comment comment = commentRepository.getByCommentId(commentId).orElseThrow(() -> new EntityNotFoundException("존재하지 않는 댓글입니다."));
+        if (!comment.getUser().getUserId().equals(user.getUserId())) {
+            throw new IllegalArgumentException("본인이 쓴 댓글만 수정할 수 있습니다.");
+        }
 
         comment.setContent(reqDto.getContent());
         commentRepository.save(comment); // 생략 가능하나, 생략하면 Auditing 기능이 동작 안함
@@ -63,9 +79,19 @@ public class CommentService {
 
     }
 
-    public ResponseDto deleteComment(Long scheduleId, Long commentId) {
-        // Schedule schedule = scheduleRepository.findByScheduleId(scheduleId).orElseThrow(() -> new EntityNotFoundException("존재하지 않는 일정입니다."));
-        Comment comment = commentRepository.getByCommentId(commentId).orElseThrow(() -> new EntityNotFoundException("존재하지 않는 댓글입니다."));
+    public ResponseDto deleteComment(Long jwtUserId, Long scheduleId, Long commentId) {
+        User user = userRepository.findById(jwtUserId).orElseThrow(
+                () -> new EntityNotFoundException("존재하지 않는 유저입니다.")
+        );
+        scheduleRepository.findByScheduleId(scheduleId).orElseThrow(
+                () -> new EntityNotFoundException("존재하지 않는 일정입니다.")
+        );
+        Comment comment = commentRepository.getByCommentId(commentId).orElseThrow(
+                () -> new EntityNotFoundException("존재하지 않는 댓글입니다.")
+        );
+        if (!comment.getUser().getUserId().equals(user.getUserId())) {
+            throw new IllegalArgumentException("본인이 쓴 댓글만 삭제할 수 있습니다.");
+        }
 
         commentRepository.deleteById(commentId);
